Create NotQuitePlatformer Scripts
Created and finished the two scripts used to run the NotQuitePlatformer game: the PlayerController.cs script goes onto the sprite and controls its movement speed and physics. The CameraFollow.cs script goes onto the camera and allows the camera to follow the sprite around the screen as it moves side to side.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>