Complete Pong scripts
Completed the PongBall.cs, PongGoal.cs, PongManager.cs, and PongPaddle.cs scripts. Moved the game to 0,0 coordinates to make the coding easier. Modified the starting y direction so the ball would stop going directly up/down. This makes sure that the ball gets to a paddle before stalling in the middle bouncing up/down. Changed the ball from a square to a circle so the bouncing is more random (with a square it would get stuck bouncing on the same y level and be boring). Finally, added the score text boxes. There is a current bug where the game won't finish after reaching the amount needed to win.