Modify bullets and fix scripts
Fixed the scripts so that the enemies properly remove health when they collide with the player. Furthermore, made it so the bullets kill the enemy once they collide.

Created a health bar in the top right corner that gets lower as more enemies hit you.

We still need to create more enemies, spawners, and a reticle. Additionally, the other bullet types need to be brought back in since they were temporarily removed for debugging.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>